requirements.txt
Flask==2.3.2
requests
rdkit-pypi
gunicorn

